+++ app/src/main/java/com/watch/aware/app/models/MonthlyData.kt
@@ -144,7 +144,7 @@ class MonthlyData {
                 }
                 "cal" -> {
                     for(steps in dteps) {
-                        stepsCnt = stepsCnt + steps.cal.toInt()
+                        stepsCnt = stepsCnt + steps.cal.toFloat()
                     }
                 }
             }
